embedder/requirements.txt
faiss-cpu==1.13.2
fastapi==0.135.3
uvicorn==0.44.0
httpx==0.28.1
aiokafka==0.13.0
